85: Upgrade all files to use app router for latest version of nextjs and turbopack
