+++ site/app/privacy-policy/page.mdx
@@ -1,6 +1,6 @@
 import React from "react";
-import PageTemplate from "../templates/Page";
-import Section from "../components/Section";
+import PageTemplate from "../../templates/Page";
+import Section from "../../components/Section";
 
 <PageTemplate metaTitle='Privacy Policy » Vibes DIY Studio'>
     <Section title='Privacy Policy'>
