+++ site/app/terms/page.mdx
@@ -1,6 +1,6 @@
 import React from 'react';
-import PageTemplate from '../templates/Page';
-import Section from '../components/Section';
+import PageTemplate from '../../templates/Page';
+import Section from '../../components/Section';
 
 <PageTemplate metaTitle='Terms & Conditions » Vibes DIY Studio'>
     <Section title='Terms & Conditions'>
