+++ site/mdx-components.tsx
@@ -1,0 +1,7 @@
+import type { MDXComponents } from "mdx/types";
+
+const components: MDXComponents = {};
+
+export function useMDXComponents(): MDXComponents {
+  return components;
+}
