+++ site/package.json
@@ -2,6 +2,7 @@
 	"name": "site",
 	"version": "0.1.0",
 	"private": true,
+	"type": "module",
 	"scripts": {
 		"dev": "next dev",
 		"build": "next build",
@@ -14,7 +15,10 @@
 		"@emailjs/browser": "^3.10.0",
 		"@emotion/react": "^11",
 		"@emotion/styled": "^11",
+		"@mdx-js/loader": "^3.1.1",
 		"@mdx-js/react": "^3.1.1",
+		"@next/mdx": "^16.1.3",
+		"@types/mdx": "^2.0.13",
 		"@types/node": "^18.11.12",
 		"@types/react": "^18.0.26",
 		"@types/react-dom": "18.0.9",
