+++ site/templates/Page/Page.module.scss
@@ -1,3 +1,17 @@
 .PageTemplate {
     min-height: 80vh;
+	h1,h2,h3,h4,h5,h6 {
+		margin-bottom: 24px;
+		font-weight: bold;
+	}
+	h2 {
+		font-size: 24px;
+	}
+	h3 {
+		font-size: 20px;
+	}
+	p {
+		margin-bottom: 16px;
+		line-height: 1.75;
+	}
 }
